requirements.txt
transformers>=4.36.0
accelerate>=0.25.0
torch>=2.0.0
langchain-community>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6
numpy>=1.24.0
pandas>=2.1.0
PyYAML>=6.0.1
gradio>=4.12.0
faiss-cpu>=1.7.4
bitsandbytes>=0.40.2
